+++ src/DbLive/Adapter/CodeItemDto.cs
@@ -4,7 +4,7 @@
 public class CodeItemDto
 {
 	public required string RelativePath { get; set; }
-	public required int ContentHash { get; set; }
+	public required long ContentHash { get; set; }
 	public required CodeItemStatus Status { get; set; }
 	public DateTime AppliedUtc { get; set; }
 	public DateTime CreatedUtc { get; set; }
