+++ src/DbLive/Adapter/MigrationItemSaveDto.cs
@@ -8,7 +8,7 @@ public class MigrationItemSaveDto
 	public required string RelativePath { get; init; }
 	public required MigrationItemType ItemType { get; init; }
 	public required MigrationItemStatus Status { get; init; }
-	public required int ContentHash { get; init; }
+	public required long ContentHash { get; init; }
 	public string? Content { get; set; }
 	public DateTime CreatedUtc { get; init; }
 }
